refactor(tests): move prove_and_verify_vm_minimal to prove_elfs_tests
