+++ prover/src/tests/prove_elfs_tests.rs
@@ -28,7 +28,7 @@ use crate::tables::trace_builder::Traces;
 use crate::test_utils::{
     E, F, collect_bitwise_lookups_from_logs, collect_bitwise_lookups_from_lt,
     collect_lt_lookups_from_logs, create_bitwise_air, create_cpu_air, create_lt_air,
-    generate_minimal_bitwise_trace, prove_and_verify_vm_minimal, run_asm_elf,
+    generate_minimal_bitwise_trace, run_asm_elf,
 };
 
 /// Alias for compatibility with existing test code.
@@ -50,6 +50,48 @@ fn collect_bitwise_lookups(
 /// Uses the FULL 2^20 row bitwise table with preprocessed commitment.
 /// Returns true if verification succeeds.
 fn prove_and_verify_vm(
+    cpu_trace: &mut stark::trace::TraceTable<F, E>,
+    bitwise_trace: &mut stark::trace::TraceTable<F, E>,
+    lt_trace: &mut stark::trace::TraceTable<F, E>,
+) -> bool {
+    let proof_options = ProofOptions::default_test_options();
+
+    let cpu_air = create_cpu_air(&proof_options);
+    let bitwise_air = create_bitwise_air(&proof_options);
+    let lt_air = create_lt_air(&proof_options);
+
+    let air_trace_pairs: Vec<(
+        &dyn AIR<Field = F, FieldExtension = E, PublicInputs = ()>,
+        _,
+        _,
+    )> = vec![
+        (&cpu_air, cpu_trace, &()),
+        (&bitwise_air, bitwise_trace, &()),
+        (&lt_air, lt_trace, &()),
+    ];
+
+    let multi_proof =
+        match Prover::multi_prove(air_trace_pairs, &mut DefaultTranscript::<E>::new(&[])) {
+            Ok(proof) => proof,
+            Err(e) => {
+                eprintln!("Prover error: {:?}", e);
+                return false;
+            }
+        };
+
+    let airs: Vec<&dyn AIR<Field = F, FieldExtension = E, PublicInputs = ()>> =
+        vec![&cpu_air, &bitwise_air, &lt_air];
+
+    Verifier::multi_verify(&airs, &multi_proof, &mut DefaultTranscript::<E>::new(&[]))
+}
+
+/// Run multi_prove and multi_verify for all VM tables with MINIMAL bitwise.
+///
+/// Used for fast tests where the bitwise table is a dummy that only contains
+/// the rows needed to balance the bus. NOT the full preprocessed table.
+///
+/// **WARNING: FOR TESTING ONLY - NOT PRODUCTION SAFE!**
+fn prove_and_verify_vm_minimal(
     cpu_trace: &mut stark::trace::TraceTable<F, E>,
     bitwise_trace: &mut stark::trace::TraceTable<F, E>,
     lt_trace: &mut stark::trace::TraceTable<F, E>,
